Some code cleanup: - refactored the particle classes to use record providers & have their own packages. - Removed a lot of commented out client config code.
